+++ build.sbt
@@ -1,4 +1,4 @@
-val scala3Version = "3.0.0-RC1"
+val scala3Version = "3.0.0"
 
 lazy val root = project
   .in(file("."))
