+++ project/plugins.sbt
@@ -1,1 +1,1 @@
-addSbtPlugin("ch.epfl.lamp" % "sbt-dotty" % "0.5.3")
+//addSbtPlugin("ch.epfl.lamp" % "sbt-dotty" % "0.5.3")
