+++ src/main/scala/syntax/CompileTime.scala
@@ -1,0 +1,13 @@
+package syntax
+
+import scala.quoted._
+
+
+inline def getPos(any: Any) = ${init('{any})}
+
+def init(expr: Expr[Any])(using Quotes) = {
+    import quotes.reflect._
+    report.error("hello",expr.asTerm.pos)
+    //error("test error",expr.asTerm.pos)
+    Expr("test")
+}
