+++ src/main/scala/syntax/SyntaxDefinition.scala
@@ -1,0 +1,5 @@
+package syntax
+
+trait SyntaxDefinition {
+    
+}
